+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add ability to pipe stdio from the node child process.
 - Update readme's github badges.
 
 ## 2.0.1